backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2
django-cors-headers==4.7.0
django-grappelli==4.0.1
djangorestframework==3.16.0
idna==3.10
requests==2.32.3
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.2
urllib3==2.4.0
